Change RustBuffer length/capacity to `u64` (#1976)

+++ uniffi_bindgen/src/bindings/swift/templates/BridgingHeaderTemplate.h
@@ -24,8 +24,8 @@
 
 typedef struct RustBuffer
 {
-    int32_t capacity;
-    int32_t len;
+    uint64_t capacity;
+    uint64_t len;
     uint8_t *_Nullable data;
 } RustBuffer;
